Move user cert generation to re-ordered cert test
